Fix `retrieve_nyms` by assigning the correct fingerprint
Due to an [update] to python-gnupg 0.3.8, `list_keys()` also returns a
`key_map` which maps keys and subkeys, and instead of always returning
dictionaries with the fingerprints of the public master keys, in certain
cases return the fingerprints corresponding to subkeys, such as passing
`secret=True` to the method.

As a consequence, nymphemeral was assigning the wrong fingerprints to
the nyms after their creation, preventing the Axolotl databases from
being loaded (searching for the wrong files) and possibly causing bugs
to everything that depended on those fingerprints.

The new `key_map` attribute was used to retrieve the correct public
master key and prevent additional searches to link keys to subkeys.

Thanks to @rxcomm for reporting and helping fix the bug.

[update]: https://bitbucket.org/vinay.sajip/python-gnupg/commits/1579ac24064c921f9f9f680a583eda99d3354790