Update the `Main Dependencies` section
Ask the user to install dependencies not installable with pip, as
nymphemeral no longer calls the package manager to install them.

Also move the `Updating` section to the bottom, since it could be
skipped by users who use pip and jump to the `Other Dependencies`
section as it is suggested by the docs.