Enable the client to generate ephemeral and hSub keys
Using the new `get_random_key` function, the creation and
configuration of a nym got simplified as the user does not need to
generate the ephemeral and hSub keys themselves.

Both the `send_create` and `send_config` methods also received minor
improvements to deal with missing parameters required for these
operations.