Enable the GUI to generate ephemeral and hSub keys
Allow the user to ask nymphemeral to automatically generate ephemeral
and hSub keys via `Checkbuttons` during nym creation and
configuration. Each `Checkbutton` is bound to the `Entry` responsible
for receiving a key manually generated by the user. By binding those
widgets, when the `Checkbutton` is checked the `Entry` is disabled
for a better UX.