Add method to retrieve a specific nym
This method is similar to both `retrieve_key` and `retrieve_nyms`. One
retrieves a specific key and the other, all the nyms. Previously,
the result of `retrieve_nyms` was being filtered manually by whatever
needed the nym. The new `retrieve_nym` method was made to meet that
need.

The `Client.start_session` method does not need to filter the nyms
anymore and the `Client.send_create` now uses the new `retrieve_nym`
method so that the client keeps reference of a `Nym` object with more
information.